in-context-information-retrieval-langchain/requirements.txt
langchain
langchain-community
langchain-google-vertexai
pypdf
python-dotenv
urllib3==1.26.18